// src/lib/db/type.ts
export interface PropertyForm {
  chooseType: string;
  houseType: string;
  Preview: string;
  Cancel: string;
  Delete: string;
  moveToDraft: string;
  pleaseResolveTheErrorsBeforeSovingToDraft: string;
  Validation: string;
  numberOfRooms: string;
  numberOfBathroom: string;
  numberOfFloors: string;
  Renovation: string;
  Furniture: string;
  Utilities: string;
  Additional: string;
  ID: string;
  "Land surface": string;
  sqM: string;
  Address: string;
  Description: string;
  mSquare: string;
  houseSurface: string;
  HouseArea: string;
  monthlyFee: string;
  dailyFee: string;
  price: string;
  enterPrice: string;
  generateCustomId: string;
  writeDescription: string;
  WriteArmDescription: string;
  AddARMDescription: string;
  AddRusDescription: string;
  WriteRusDescription: string;
  AddEnTitle: string;
  WriteEnTitle: string;
  ProvinceStateCity: string;
  WriteEnDescription: string;
  AddEnDescription: string;
  Province: string;
  District: string;
  RewriteTitle: string;
  Title: string;
  AddArmtitle: string;
  WriteRusTitle: string;
  ConstructionType: string;
  BuildingType: string;
  "Floor area sq/m": string;
  floor: string;
  Area: string;
  "House Area": string;
  Significance: string;
  "Cancel Action": string;
  "House type": string;
  "Building type": string;
  "Construction type": string;
  "Do you want to cancel this action or move it to draft?": string;
  "This property already exists in your account.": string;
  "Duplicate Error": string;
  Post: string;
  "Post an Ad": string;
  Back: string;
  "Choose category": string;
  Info: string;
  "title.required": string;
  "description.required": string;
  "price.required": string;
  "location.required": string;
  "images.required": string;
  "details.bathroomRequired": string;
  "details.bedroomRequired": string;
  "details.areaRequired": string;
  "details.totalAreasRequired": string;
  "details.utilityRequired": string;
  "details.amenityRequired": string;
  "details.SignificanceRequired": string;
  "houseType.required": string;
  "buildingType.required": string;
  "ConstructionType.required": string;
  "renovation.required": string;
  "details.zoningTypeRequired": string;
  "details.businessTypeRequired": string;
  "Furniture type is required": string;
  "Property Updated and moved to draft": string;
  "Property Updated successfully": string;
  Success: string;
  Error: string;
  "Fill in": string;
  "Best offer?": string;
  "Drop or upload photo (max": string;
  max: string;
  photos: string;
  "You can only upload up to": string;
  // "Площадь дома (кв/м)": string;
  House: string;
  "Enter type 3/16X format": string;
  Only: string;
  imageUploadError: string;
  isContractBased: string;
}